Fixed Broken HD URL(s)
Due to 'whvl' not having HD content the IF statement set every channel
after it in the list from HD to HIGH quality. The If statement has been
resolved and whvl is only channel to be reduced from HD to HIGH when
preferences is set to HD; as well as first steps to reducing API/JSON
time.